Get some ad hoc bumper finding code working
"Ad hoc" as in "it recognizes the red and blue bumpers we used in
Steamworks when they're on top of each other". I need to test it more,
but I can't do it today because using my computer's built-in webcam for
GRIP development is a PITA and the only spare webcam we have requires me
to take an Ethernet-to-USB adapter with me.